chore(Residents): added validators to  resident module
